major refactor added KeepMigrationsForDays transfer configuration setting
