+++ Source/ACE.Server/Managers/TransferManager/Enums/ImportAndMigrateFailiureReason.cs
@@ -1,0 +1,36 @@
+namespace ACE.Server.Managers.TransferManager.Enums
+{
+    public enum ImportAndMigrateFailiureReason
+    {
+        None,
+        Unknown,
+        OperationNotAllowed,
+        UnverifiedSourceServerNotAllowed,
+        UnverifiedSourceServerDoesntAllowMigrate,
+        VerifiedSourceServerNotAllowed,
+        VerifiedMigrationSourceThumbprintMismatch,
+        PackInfoNotFound,
+        CannotContactSourceServer,
+        ProtocolError,
+        SourceServerRejectedRequest,
+        CharacterAlreadyPresent,
+        WrongPackageType,
+        PackageTypeNotAllowed,
+        NameIsUnavailable,
+        NameIsNaughty,
+        NameContainsInvalidCharacters,
+        NameTooShortOrTooLong,
+        NoCharacterSlotsAvailable,
+        FoundMoreThanOneCharacter,
+        CannotFindCharacter,
+        MalformedCharacterData,
+        AddCharacterFailed,
+        CharacterPackageSignatureInvalid,
+        SignatureMismatch,
+        SignatureInvalid,
+        MigrationCheckFailed,
+        NonceInvalid,
+        PackageUnsigned,
+        CookieAlreadyUsed
+    }
+}
