+++ Source/ACE.Server/Managers/TransferManager/Enums/MigrationCloseType.cs
@@ -1,0 +1,8 @@
+namespace ACE.Server.Managers.TransferManager.Enums
+{
+    public enum MigrationCloseType
+    {
+        Cancel,
+        Download
+    }
+}
