+++ Source/ACE.Server/Managers/TransferManager/Enums/MigrationReadyStatus.cs
@@ -1,0 +1,15 @@
+namespace ACE.Server.Managers.TransferManager.Enums
+{
+    public enum MigrationReadyStatus
+    {
+        Unknown,
+        Ready,
+        NonExistant,
+        NotMigration,
+        AlreadyDownloaded,
+        AlreadyCancelled,
+        CharNotFound,
+        InvalidCharState,
+        PackageFileMissing
+    }
+}
