+++ Source/ACE.Server/Managers/TransferManager/Enums/PackageType.cs
@@ -1,0 +1,8 @@
+namespace ACE.Server.Managers.TransferManager.Enums
+{
+    public enum PackageType
+    {
+        Migrate,
+        Backup
+    }
+}
