+++ Source/ACE.Server/Managers/TransferManager/MigrateCloseResult.cs
@@ -1,0 +1,8 @@
+namespace ACE.Server.Managers.TransferManager
+{
+    public class MigrateCloseResult
+    {
+        public bool Success { get; set; } = false;
+        public byte[] SnapshotPackage { get; set; }
+    }
+}
