+++ Source/ACE.Server/Managers/TransferManager/Responses/MigrationCheckRequestModel.cs
@@ -1,0 +1,8 @@
+namespace ACE.Server.Managers.TransferManager.Responses
+{
+    public class MigrationCheckRequestModel
+    {
+        public string Nonce { get; set; }
+        public string Cookie { get; set; }
+    }
+}
