+++ Source/ACE.Server/Managers/TransferManager/Responses/SignedTransferManagerMigrationCheckResponseModel.cs
@@ -1,0 +1,9 @@
+namespace ACE.Server.Managers.TransferManager.Responses
+{
+    public class SignedMigrationCheckResponseModel
+    {
+        public MigrationCheckResponseModel Result { get; set; }
+        public byte[] Signature { get; set; }
+        public byte[] Signer { get; set; }
+    }
+}
