+++ Source/ACE.WebApiServer/Model/Character/Migration/Cancel.cs
@@ -1,4 +1,4 @@
-using ACE.Server.Managers;
+using ACE.Server.Managers.TransferManager;
 using FluentValidation;
 
 namespace ACE.WebApiServer.Model.Character.Migration
@@ -14,12 +14,12 @@ namespace ACE.WebApiServer.Model.Character.Migration
             RuleFor(request => request.Cookie).NotEmpty().WithMessage("You must specify the character migration cookie.");
             RuleFor(request => request.Cookie).Custom((str, _) =>
             {
-                if (TransferManager.StringContainsInvalidChars(TransferManager.CookieChars, str))
+                if (TransferManagerUtil.StringContainsInvalidChars(TransferManagerConstants.CookieChars, str))
                 {
                     _.AddFailure("The cookie contains invalid characters.");
                 }
             });
-            RuleFor(request => request.Cookie).Length(TransferManager.CookieLength).WithMessage($"Cookie must be {TransferManager.CookieLength} characters in length.");
+            RuleFor(request => request.Cookie).Length(TransferManagerConstants.CookieLength).WithMessage($"Cookie must be {TransferManagerConstants.CookieLength} characters in length.");
         }
     }
     public class CharacterMigrationCancelResponseModel
