+++ Source/ACE.WebApiServer/Model/Character/Migration/Download.cs
@@ -1,4 +1,5 @@
-using ACE.Server.Managers;
+using ACE.Server.Managers.TransferManager;
+using ACE.Server.Managers.TransferManager.Responses;
 using FluentValidation;
 
 namespace ACE.WebApiServer.Model.Character.Migration
@@ -14,13 +15,13 @@ namespace ACE.WebApiServer.Model.Character.Migration
             RuleFor(request => request.Cookie).NotEmpty().WithMessage("You must specify the character migration cookie.");
             RuleFor(request => request.Cookie).Custom((str, _) =>
             {
-                if (TransferManager.StringContainsInvalidChars(TransferManager.CookieChars, str))
+                if (TransferManagerUtil.StringContainsInvalidChars(TransferManagerConstants.CookieChars, str))
                 {
                     _.AddFailure("The cookie contains invalid characters.");
                 }
             });
-            RuleFor(request => request.Cookie).Length(TransferManager.CookieLength).WithMessage($"Cookie must be {TransferManager.CookieLength} characters in length.");
+            RuleFor(request => request.Cookie).Length(TransferManagerConstants.CookieLength).WithMessage($"Cookie must be {TransferManagerConstants.CookieLength} characters in length.");
         }
     }
-    public class CharacterMigrationDownloadResponseModel : TransferManager.TransferManagerCharacterMigrationDownloadResponseModel { }
+    public class WebApiCharacterMigrationDownloadResponseModel : CharacterMigrationDownloadResponseModel { }
 }
